feat: refactor Party and PartyException classes; add unit tests for Party and Vibe
