+++ lesson_16/objects/objects_app/src/test/java/com/codedifferently/lesson16/Lesson16Test.java
@@ -5,8 +5,6 @@ import static org.assertj.core.api.Assertions.assertThat;
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
 
-import com.codedifferently.lesson16.Lesson16;
-
 @SpringBootTest(useMainMethod = SpringBootTest.UseMainMethod.WHEN_AVAILABLE)
 class Lesson16Test {
