+++ lesson_16/objects/objects_app/src/test/java/com/codedifferently/lesson16/ezranyabuti/PartyExceptionTest.java
@@ -1,12 +1,10 @@
-package test.java.com.codedifferently.lesson16.ezranyabuti;
+package com.codedifferently.lesson16.ezranyabuti;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
-import com.codedifferently.lesson16.ezranyabuti.PartyException;
 import org.junit.jupiter.api.Test;
 
 public class PartyExceptionTest {
-  // Test for the constructor
   @Test
   void testConstructor() {
     String expectedMessage = "Party is not private";
